pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sereto"
version = "0.0.1"
description = "Security Reporting Tool."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
  { name = "Tomáš Stefan", email = "t@s4n.cz" },
  { name = "NN Management Services, s.r.o." },
]
keywords = [
  "automate",
  "markdown",
  "pandocfilters",
  "pdf",
  "penetration testing",
  "pentest",
  "report",
  "tex",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Framework :: Pydantic :: 2",
  "Framework :: Pydantic",
  "Framework :: Pytest",
  "Framework :: tox",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python",
  "Topic :: Security",
  "Topic :: Text Processing :: Markup :: LaTeX",
  "Topic :: Text Processing :: Markup :: Markdown",
  "Topic :: Text Processing :: Markup",
  "Topic :: Text Processing",
  "Typing :: Typed",
]
dependencies = [
  "click",
  "jinja2",
  "matplotlib",
  "pathspec",
  "pydantic_settings",
  "pydantic>=2.6",
  "pydantic[email]",
  "pypdf",
  "python-frontmatter",
  "rich",
  "ruamel.yaml",
  "semver",
  "toml",
  "Unidecode",
]

[project.urls]
Changelog = "https://github.com/tstdin/SeReTo/blob/main/CHANGELOG.md"
# Documentation = "TODO"
Homepage = "https://github.com/tstdin/SeReTo"
Source = "https://github.com/tstdin/SeReTo"

[project.scripts]
sereto = "sereto.cli.cli:cli"

[tool.black]
line-length = 119

[tool.flake8]
max-line-length = 119

[tool.isort]
profile = "black"

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = [
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
  # pycodestyle
  "E",
  "W",
  # pydocstyle
  # "D",  # TODO
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
]
